requirements.txt
feedparser>=6.0.11
python-dateutil>=2.9.0
python-dotenv>=1.0.1
rapidfuzz>=3.9.6
requests>=2.32.3
# 선택: Gemini 무료 API (Tier 1 속보 검증용)
google-generativeai>=0.8.0
